extract and move LeipzigRoutingStrategyProvider
